+++ tests/TestCase.php
@@ -101,7 +101,6 @@ abstract class TestCase extends \Orchestra\Testbench\TestCase
 
         $app['config']->set('statamic.api.enabled', true);
         $app['config']->set('statamic.graphql.enabled', true);
-
         $app['config']->set('statamic.editions.pro', true);
 
         $app['config']->set('cache.stores.outpost', [